huge refactor and add slash commands
