+++ src/commands/slash/mod.rs
@@ -1,0 +1,1 @@
+pub mod ping;
