+++ src/commands/slash/ping.rs
@@ -1,0 +1,23 @@
+use serenity::{
+    builder::CreateApplicationCommand,
+    model::prelude::interaction::{
+        application_command::ApplicationCommandInteraction, InteractionResponseType,
+    },
+    prelude::Context,
+};
+
+pub async fn run(ctx: &Context, command: &ApplicationCommandInteraction) {
+    if let Err(e) = command
+        .create_interaction_response(&ctx.http, |resp| {
+            resp.kind(InteractionResponseType::ChannelMessageWithSource)
+                .interaction_response_data(|msg| msg.content("Pong baby!"))
+        })
+        .await
+    {
+        println!("unable to respond to slash command: {}", e)
+    }
+}
+
+pub fn register(command: &mut CreateApplicationCommand) -> &mut CreateApplicationCommand {
+    command.name("ping").description("A ping command")
+}
