+++ src/utils/syllables.rs
@@ -1,9 +1,4 @@
-use std::collections::HashMap;
-use std::fs;
-use std::fs::File;
-use std::io::{self, BufRead};
-use std::path::Path;
-
+#[allow(dead_code)]
 pub fn count_syllables(word: &str) -> u8 {
     let mut count = 0;
     let mut prev_was_vowel = false;
@@ -21,34 +16,6 @@ pub fn count_syllables(word: &str) -> u8 {
         count -= 1;
     }
     count
-}
-
-pub fn read_dir(dir: String) -> HashMap<String, Vec<String>> {
-    let paths = fs::read_dir(dir).unwrap();
-    let mut map: HashMap<String, Vec<String>> = HashMap::new();
-    for path in paths {
-        let mut cur: Vec<String> = Vec::new();
-        let curpath = path.unwrap().path();
-        let filename = curpath.file_name().unwrap().to_str().unwrap().to_string();
-        let filetype = filename[..(filename.len() - 4)].to_string();
-        if let Ok(lines) = read_lines(curpath) {
-            for line in lines.flatten() {
-                if !line.is_empty() {
-                    cur.push(line);
-                }
-            }
-            map.insert(filetype, cur);
-        }
-    }
-    map
-}
-
-fn read_lines<P>(filename: P) -> io::Result<io::Lines<io::BufReader<File>>>
-where
-    P: AsRef<Path>,
-{
-    let file = File::open(filename)?;
-    Ok(io::BufReader::new(file).lines())
 }
 
 #[cfg(test)]
